requirements.txt
requests>=2.32.0,<3
mysql-connector-python>=9.1.0,<10
kafka-python-ng>=2.2.3,<3
